requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/requirements.in
#
asgiref==3.8.1
    # via django
dj-database-url==2.1.0
    # via -r requirements/requirements.in
django==5.0.6
    # via
    #   -r requirements/requirements.in
    #   dj-database-url
    #   djangorestframework
djangorestframework==3.15.1
    # via -r requirements/requirements.in
psycopg2==2.9.9
    # via -r requirements/requirements.in
python-decouple==3.8
    # via -r requirements/requirements.in
sqlparse==0.5.0
    # via django
typing-extensions==4.11.0
    # via
    #   asgiref
    #   dj-database-url
